Removing compiler warnings in MSVC 2005 build
git-svn-id: http://svn.code.sf.net/p/freeglut/code/trunk/freeglut/freeglut@725 7f0cb862-5218-0410-a997-914c9d46530a

+++ freeglut/src/freeglut_init.c
@@ -720,7 +720,7 @@ void FGAPIENTRY glutInitDisplayString( const char* displayMode )
      * delimited by blanks or tabs.
      */
     char *token ;
-    int len = strlen ( displayMode );
+    size_t len = strlen ( displayMode );
     char *buffer = (char *)malloc ( (len+1) * sizeof(char) );
     memcpy ( buffer, displayMode, len );
     buffer[len] = '\0';
